Extract @optimitron/tracking package and mount the dFDA MCP server
Move the 10 personal tracking MCP tools (core logic, tool definitions,
dispatch — ~2,400 lines) from packages/web/src/lib/mcp-server.ts into the
new packages/tracking, behavior unchanged, with the Prisma client injected
by each host. optimitron.com keeps serving the same tools via the package;
dfda.earth/api/mcp now serves them too as an OAuth resource server
(optimitron.com stays the sole authorization server; RFC 9728 metadata
route added). Introduce shared MCP_TOKEN_SECRET (NEXTAUTH_SECRET fallback)
so satellite Vercel projects can verify the same Bearer JWTs.

Amend doctrine docs per the 2026-08-13 architecture decision on task
optimitron:dev:dfda-tracking-mcp-server-split: ROADMAP Won't entry,
PRD 9.3 REST exception, FEATURES evidence paths, SYSTEM_MAP, AGENTS
dependency graph.
