+++ .env.example
@@ -12,6 +12,9 @@ NEXTAUTH_URL=http://localhost:3001
 # MCP_OAUTH_ISSUER=https://optimitron.example.com
 # Generate locally, for example: `openssl rand -base64 32`
 NEXTAUTH_SECRET=replace-this-with-a-long-random-secret
+# Optional dedicated MCP JWT signing secret, shared with satellite resource
+# servers (apps/dfda). Falls back to NEXTAUTH_SECRET when unset.
+MCP_TOKEN_SECRET=
 
 # Email / auth providers for packages/web
 # Resend API keys: https://resend.com/docs/dashboard/api-keys/introduction
