+++ apps/dfda/.env.example
@@ -1,6 +1,10 @@
 # Database and authentication
 DATABASE_URL="postgresql://user:password@host-pooler/database?sslmode=require"
 NEXTAUTH_SECRET="generate-a-random-32-byte-secret"
+# Must equal the optimitron web project's MCP token secret so dfda.earth/api/mcp
+# can verify Bearer tokens issued by optimitron.com. Locally the shared root .env
+# NEXTAUTH_SECRET fallback covers it.
+MCP_TOKEN_SECRET=""
 NEXTAUTH_URL="http://localhost:3011"
 
 # Optional sign-in and email providers
