+++ apps/dfda/package.json
@@ -18,6 +18,7 @@
     "@google/genai": "^1.50.1",
     "@hello-pangea/dnd": "^18.0.1",
     "@hookform/resolvers": "^3.10.0",
+    "@modelcontextprotocol/sdk": "^1.29.0",
     "@neondatabase/serverless": "^1.1.0",
     "@next/third-parties": "^15.5.15",
     "@optimitron/data": "workspace:^",
@@ -25,6 +26,7 @@
     "@optimitron/impact-params": "workspace:^",
     "@optimitron/neobrutalist-ui": "workspace:^",
     "@optimitron/site-kit": "workspace:^",
+    "@optimitron/tracking": "workspace:^",
     "@prisma/adapter-pg": "^7.8.0",
     "@radix-ui/react-accordion": "1.2.2",
     "@radix-ui/react-alert-dialog": "1.1.4",
@@ -71,6 +73,7 @@
     "geist": "latest",
     "glob": "^13.0.6",
     "input-otp": "1.4.1",
+    "jose": "^6.2.2",
     "katex": "^0.16.45",
     "lucide-react": "^0.454.0",
     "nanoid": "5.1.6",
